CAMEL-15108: Extract the implementation of the JtaTransactionPolicy from camel-cdi to camel-jta
