+++ components/camel-cdi/src/main/java/org/apache/camel/cdi/xml/ErrorHandlerType.java
@@ -23,7 +23,7 @@ import org.apache.camel.builder.DeadLetterChannelBuilder;
 import org.apache.camel.builder.DefaultErrorHandlerBuilder;
 import org.apache.camel.builder.ErrorHandlerBuilder;
 import org.apache.camel.builder.NoErrorHandlerBuilder;
-import org.apache.camel.cdi.transaction.JtaTransactionErrorHandlerBuilder;
+import org.apache.camel.jta.JtaTransactionErrorHandlerBuilder;
 
 /**
  * Used to configure the errorHandler type
