+++ components/camel-jta/src/main/java/org/apache/camel/jta/JtaTransactionErrorHandlerBuilder.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.cdi.transaction;
+package org.apache.camel.jta;
 
 import org.apache.camel.LoggingLevel;
 import org.apache.camel.builder.DefaultErrorHandlerBuilder;
